@@ -18,20 +18,23 @@
         }
     ],
     "require": {
-        "php": "^7.1",
+        "php": "^7.4",
         "php-http/httplug": "^1.0 || ^2.0",
         "php-http/multipart-stream-builder": "^1.0",
         "php-http/client-common": "^2.0",
         "php-http/discovery": "^1.0"
     },
     "require-dev": {
-        "phpunit/phpunit": "^7.4",
+        "phpunit/phpunit": "^9.5",
         "php-http/message-factory": "^1.0",
         "php-http/curl-client": "^2.0",
         "php-http/message": "^1.0",
         "symfony/http-client": "^5.0",
         "nyholm/nsa": "^1.1",
-        "nyholm/psr7": "^1.3"
+        "nyholm/psr7": "^1.3",
+        "symfony/var-dumper": "^5.3",
+        "sylius-labs/coding-standard": "^4.1",
+        "php-vcr/php-vcr": "^1.5"
     },
     "autoload": {
         "psr-4": {
@@ -40,11 +43,13 @@
     },
     "autoload-dev": {
         "psr-4": {
-            "FAPI\\Boilerplate\\Tests\\": "tests/"
+            "Tests\\": "tests/"
         }
     },
     "scripts": {
         "test": "vendor/bin/phpunit",
-        "test-ci": "vendor/bin/phpunit --coverage-text --coverage-clover=build/coverage.xml"
+        "test-ci": "vendor/bin/phpunit --coverage-text --coverage-clover=build/coverage.xml",
+        "ecs:fixer": "vendor/bin/ecs check src tests",
+        "ecs:fixer:fix": "vendor/bin/ecs check src tests --fix"
     }
 }
