@@ -10,38 +10,19 @@ use Http\Client\HttpClient;
 /**
  * Class that gets access tokens.
  *
- * @author Tobias Nyholm <tobias.nyholm@gmail.com>
- *
  * @internal
  */
 final class Authenticator
 {
-    /**
-     * @var RequestBuilder
-     */
-    private $requestBuilder;
+    private RequestBuilder $requestBuilder;
 
-    /**
-     * @var HttpClient
-     */
-    private $httpClient;
+    private HttpClient $httpClient;
 
-    /**
-     * @var string|null
-     */
-    private $accessToken;
+    private ?string $accessToken;
 
-    /**
-     * @var string
-     */
-    private $clientId;
+    private string $accessMode = 'shop';
 
-    /**
-     * @var string
-     */
-    private $clientSecret;
-
-    public function __construct(RequestBuilder $requestBuilder, HttpClient $httpClient, string $clientId, string $clientSecret)
+    public function __construct(RequestBuilder $requestBuilder, HttpClient $httpClient, string $clientId = null, string $clientSecret = null)
     {
         $this->requestBuilder = $requestBuilder;
         $this->httpClient = $httpClient;
@@ -49,46 +30,23 @@ final class Authenticator
         $this->clientSecret = $clientSecret;
     }
 
-    public function createAccessToken(string $username, string $password): ?string
+    public function createAccessToken(string $username, string $password, $access = 'shop'): ?string
     {
-        $request = $this->requestBuilder->create('POST', '/api/oauth/v2/token', [
-            'Content-type' => 'application/x-www-form-urlencoded',
-        ], \http_build_query([
-            'client_id' => $this->clientId,
-            'client_secret' => $this->clientSecret,
-            'grant_type' => 'password',
-            'username' => $username,
-            'password' => $password,
-        ]));
+        $this->accessMode = $access;
+        $body = ['email' => $username, 'password' => $password];
+        $path = sprintf('/api/v2/%s/authentication-token', $access);
+        $request = $this->requestBuilder->create('POST', $path, [
+            'accept' => 'application/json',
+            'Content-type' => 'application/json',
+        ], json_encode($body));
 
         $response = $this->httpClient->sendRequest($request);
+
         if (200 !== $response->getStatusCode()) {
             return null;
         }
 
-        $this->accessToken = $response->getBody()->__toString();
-
-        return $this->accessToken;
-    }
-
-    public function refreshAccessToken(string $accessToken, string $refreshToken): ?string
-    {
-        $request = $this->requestBuilder->create('POST', '/api/oauth/v2/token', [
-            'Authorization' => \sprintf('Bearer %s', $accessToken),
-            'Content-type' => 'application/x-www-form-urlencoded',
-        ], \http_build_query([
-            'refresh_token' => $refreshToken,
-            'grant_type' => 'refresh_token',
-            'client_id' => $this->clientId,
-            'client_secret' => $this->clientSecret,
-        ]));
-
-        $response = $this->httpClient->sendRequest($request);
-        if (200 !== $response->getStatusCode()) {
-            return null;
-        }
-
-        $this->accessToken = $response->getBody()->__toString();
+        $this->accessToken = json_decode($response->getBody()->__toString(), true)['token'];
 
         return $this->accessToken;
     }
@@ -96,5 +54,10 @@ final class Authenticator
     public function getAccessToken(): ?string
     {
         return $this->accessToken;
+    }
+
+    public function getAccessMode(): ?string
+    {
+        return $this->accessMode;
     }
 }
