@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 /*
  * This software may be modified and distributed under the terms
  * of the MIT license. See the LICENSE file for details.
@@ -9,8 +11,6 @@ namespace FAPI\Sylius;
 
 /**
  * All our exception implements this interface.
- *
- * @author Tobias Nyholm <tobias.nyholm@gmail.com>
  */
 interface Exception
 {
