@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 /*
  * This software may be modified and distributed under the terms
  * of the MIT license. See the LICENSE file for details.
@@ -9,9 +11,6 @@ namespace FAPI\Sylius\Exception;
 
 use FAPI\Sylius\Exception;
 
-/**
- * @author Márk Sági-Kazár <mark.sagikazar@gmail.com>
- */
 interface DomainException extends Exception
 {
 }
