@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 /*
  * This software may be modified and distributed under the terms
  * of the MIT license. See the LICENSE file for details.
@@ -7,9 +9,6 @@
 
 namespace FAPI\Sylius\Model;
 
-/**
- * @author Tobias Nyholm <tobias.nyholm@gmail.com>
- */
 interface CreatableFromArray
 {
     /**
